+++ core/src/backend/matrix/types.rs
@@ -1,9 +1,18 @@
 use super::prelude::*;
-use crate::base::RegistryState;
+use crate::base::StoreState;
 use crate::utils::prelude::*;
 
 #[derive(Serialize, Clone, Debug, Default, Deserialize, EventContent)]
 #[ruma_event(type = "m.virto.apps", kind = GlobalAccountData)]
 pub struct MatrixAppsStateContent {
-    pub apps: RegistryState,
+    pub apps: StoreState,
 }
+
+
+#[async_trait]
+pub trait MatrixSuperClient {
+    fn client() -> Client;
+    async fn wait_until_sync() -> ();
+}
+
+
