--- core/src/base/process/mod.rs
@@ -1,3 +1,0 @@
-mod process;
-
-pub use process::*;
