+++ core/src/base/registry.rs
@@ -1,5 +1,6 @@
-use crate::utils::prelude::*;
+use crate::AppLoader;
 use crate::AppInfo;
+use crate::utils::prelude::*;
 
 #[derive(Debug)]
 pub enum AppRegistryError {
@@ -9,18 +10,33 @@ pub enum AppRegistryError {
     CantUninstall(String),
 }
 
+
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct AppMetadata {
     pub app_info: AppInfo,
     pub channel_id: String,
 }
 
-pub type RegistryResult<T> = Result<T, AppRegistryError>;
-pub type RegistryState = HashMap<String, AppMetadata>;
+pub type StoreResult<T> = Result<T, AppRegistryError>;
+pub type StoreState = HashMap<String, AppMetadata>;
 
-pub trait Registry {
-    async fn is_registered(&self, id: &str) -> RegistryResult<bool>;
-    async fn add(&self, app_info: &AppInfo) -> RegistryResult<RegistryState>;
-    async fn remove(&self, id: &AppInfo) -> RegistryResult<RegistryState>;
-    async fn list_apps(&self) -> RegistryResult<Vec<AppInfo>>;
+pub trait Store {
+    async fn is_registered(&self, id: &str) -> StoreResult<bool>;
+    async fn add(&self, app_info: &AppInfo) -> StoreResult<StoreState>;
+    async fn remove(&self, id: &AppInfo) -> StoreResult<StoreState>;
+    async fn list_apps(&self) -> StoreResult<Vec<AppInfo>>;
 }
+
+
+pub type RegistryResult<T> = Result<T, AppRegistryError>;
+
+pub enum AppRegisteredLoader {
+    NotInstalled
+}
+
+#[async_trait]
+pub trait AppRegistry: Sync + Send  {
+    async fn install<'r>(&self, loader: &Box<dyn AppLoader<'r>>) -> RegistryResult<()>;
+    async fn get_loader<'r>(&self, id: &str) -> Option<&Box<dyn AppLoader<'r>>>;
+}
+
