--- core/src/base/supervisor/mod.rs
@@ -1,3 +1,0 @@
-mod supervisor;
-
-pub use supervisor::*;
