+++ core/tests/backend/mod.rs
@@ -1,0 +1,1 @@
+mod backend_test;
