+++ core/tests/fixtures/mod.rs
@@ -1,0 +1,3 @@
+
+pub mod app_mock;
+
