+++ core/tests/integration/mod.rs
@@ -1,0 +1,2 @@
+pub mod app_test;
+pub mod process_test;
