+++ core/tests/mod.rs
@@ -1,0 +1,12 @@
+
+#[cfg(test)]
+pub mod fixtures;
+
+#[cfg(test)]
+pub mod integration;
+
+// mod process_test;
+// mod sdk_test;
+// mod supervisor_test;
+
+// mod backend;
